media_codec: groupe AccountCodecInfo and SystemCodecInfo
This distinction between informations from the backend and user
preferences on the codec is not really useful and introduces some
useless complexity.

GitLab: #804

+++ src/media/media_encoder.cpp
@@ -492,7 +492,7 @@ MediaEncoder::encode(AVFrame* frame, int streamIdx)
         }
     }
     int ret = 0;
-    if (streamIdx >= encoders_.size())
+    if (static_cast<size_t>(streamIdx) >= encoders_.size())
         return -1;
     AVCodecContext* encoderCtx = encoders_[streamIdx];
     AVPacket pkt;
